chore: update plan detail page
