+++ frontend/src/components/Plan/components/PlanDataTable/PlanDataTable.vue
@@ -20,7 +20,6 @@ import { NPerformantEllipsis, NDataTable } from "naive-ui";
 import { computed, ref } from "vue";
 import { useI18n } from "vue-i18n";
 import { useRouter } from "vue-router";
-import { BBAvatar } from "@/bbkit";
 import { ProjectNameCell } from "@/components/v2/Model/DatabaseV1Table/cells";
 import { PROJECT_V1_ROUTE_PLAN_DETAIL } from "@/router/dashboard/projectV1";
 import type { ComposedPlan } from "@/types/v1/issue/plan";
@@ -63,19 +62,6 @@ const columnList = computed((): DataTableColumn<ComposedPlan>[] => {
       hide: !props.showProject,
       render: (plan) => (
         <ProjectNameCell project={plan.projectEntity} mode={"ALL_SHORT"} />
-      ),
-    },
-    {
-      key: "creator",
-      resizable: true,
-      width: 150,
-      title: t("issue.table.creator"),
-      hide: !showExtendedColumns.value,
-      render: (plan) => (
-        <div class="flex flex-row items-center overflow-hidden gap-x-2">
-          <BBAvatar size="SMALL" username={plan.creatorEntity.title} />
-          <span class="truncate">{plan.creatorEntity.title}</span>
-        </div>
       ),
     },
     {
